feat: Add initial actor sheets, item organization, condition synchronization, dice rolling, and system configuration.
